requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-telegram-bot==20.7
apscheduler==3.10.4
httpx==0.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
